# koicare/koi_swagger.yaml
openapi: 3.0.1
info:
  title: Koi Care API
  description: API documentation for your web application
  version: 1.0.0
servers:
  - url: "http://localhost:80"
    description: Production server
paths:
...
  /koi:
    get:
      summary: Get all koi
      tags:
        - Koi
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of all koi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KoiWithOwner"

    post:
      summary: Create a new koi
      tags:
        - Koi
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                image:
                  type: string
                body_shape:
                  type: string
                age:
                  type: integer
                size:
                  type: number
                  format: float
                weight:
                  type: number
                  format: float
                gender:
                  type: string
                  enum: [male, female]
                breed:
                  type: string
                origin:
                  type: string
                pond_id:
                  type: integer
              required:
                - name
                - image
                - body_shape
                - age
                - size
                - weight
                - gender
                - breed
                - origin
                - pond_id
      responses:
        "201":
          description: Koi created successfully
        "400":
          description: Bad Request
        "500":
          description: Internal server error
  /koi/{id}:
    get:
      summary: Get koi by ID
      tags:
        - Koi
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi
      responses:
        "200":
          description: Koi retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KoiWithOwner"
        "404":
          description: Koi not found
        "500":
          description: Internal server error

    put:
      summary: Update koi by ID
      tags:
        - Koi
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                body_shape:
                  type: string
                age:
                  type: integer
                size:
                  type: number
                  format: float
                weight:
                  type: number
                  format: float
                pond_id:
                  type: integer
      responses:
        "200":
          description: Koi updated successfully
        "404":
          description: Koi not found
        "500":
          description: Internal server error

    delete:
      summary: Delete koi by ID
      tags:
        - Koi
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi
      responses:
        "200":
          description: Koi deleted successfully
        "404":
          description: Koi not found
        "500":
          description: Internal server error
...
  /koi/{id}/food:
    get:
      summary: Get food calculation by koi ID
      tags:
        - Koi
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi
      responses:
        "200":
          description: Koi retrieved successfully with food calculation
          content:
            application/json:
              schema:
                type: object
                properties:
                  food_required_kg_per_day:
                    type: number
                    format: float
        "404":
          description: Koi not found
        "500":
          description: Internal server error
...
  /order:
    get:
      summary: Get all orders
      tags:
        - Order
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    order_date:
                      type: string
                      format: date-time
                    user_id:
                      type: integer
                    status:
                      type: string
                      enum: [pending, processing, shipped, delivered, cancelled]
        "500":
          description: Internal server error
    post:
      summary: Create a new order
      tags:
        - Order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderItems:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: integer
                        description: ID of the product
                      quantity:
                        type: integer
                        description: Quantity of the product
                      price:
                        type: number
                        format: float
                        description: Price of the product
                    required:
                      - product_id
                      - quantity
                      - price
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  orderId:
                    type: integer
        "400":
          description: Bad Request (e.g., invalid input)
        "500":
          description: Internal server error
  /order/{id}:
    get:
      summary: Get order by ID
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the order
      responses:
        "200":
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties: # Removed total_amount
                  order_id:
                    type: integer
                  user_id:
                    type: integer
                  order_date:
                    type: string
                    format: date-time
                  status:
                    type: string
                    enum: [pending, processing, shipped, delivered, cancelled]
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        product_id:
                          type: integer
                        name:
                          type: string
                        price:
                          type: number
                          format: float
                        quantity:
                          type: integer
        "404":
          description: Order not found
        "500":
          description: Internal server error

    put:
      summary: Update order by ID
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderItems:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: integer
                      quantity:
                        type: integer
                      price:
                        type: number
                        format: float
                status:
                  type: string
                  enum: [pending, processing, shipped, delivered, cancelled]
      responses:
        "200":
          description: Order updated successfully
        "404":
          description: Order not found
        "500":
          description: Internal server error

    delete:
      summary: Delete order by ID
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the order
      responses:
        "200":
          description: Order deleted successfully
        "404":
          description: Order not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Koi:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: string
        body_shape:
          type: string
        age:
          type: integer
        size:
          type: number
          format: float
        weight:
          type: number
          format: float
        gender:
          type: string
        breed:
          type: string
        origin:
          type: string
        pond_id:
          type: integer
    KoiWithOwner:
      allOf:
        - $ref: "#/components/schemas/Koi"
        - type: object
          properties:
            owner:
              type: string
              description: Username of the koi's owner
...